Struct rsp2_structure::bonds::FracBonds

pub struct FracBonds { /* fields omitted */ }

Bond data in a more widely-reusable form than CartBonds.

Represents each bond as a composite of the following data:

The source site.

The target site.

A lattice vector that distinguishes different images of the target.

The lattice vector is chosen such that:

bond_cart_vector == carts[to] - carts[from] + image_diff * lattice

Considering the data embedded in this representation, any of the following actions will invalidate the bonds:

removal, addition, or reordering of sites

replacing a coordinate with an image (e.g. remapping sites into the unit cell)

unimodular transformations of the lattice (these change the images)

The following actions are okay:

motion of atoms, even if they cross cell boundaries, (so long as relative distances do not change enough that the bond graph would need to change)

applying a cartesian operator to the lattice (e.g. rotating or scaling the entire structure while preserving frac coords)

Implementations

impl FracBonds

pub fn to_periodic_graph(&self) -> PeriodicGraph

Constructs a finite graph representing bonds per unitcell.

This is basically just a different representation of FracBonds, which is better equipped for certain types of lookup and modification. (It’s terrible for most graph algorithms, though)

Properties

Node indices will match site indices.

The graph will be a directed multi-graph.

Parallel edges will each represent a different image_diff, stored as the edge weight.

impl FracBonds

pub fn from_iter(
num_atoms: usize,
iter: impl IntoIterator<Item = FracBond>
) -> Self

impl FracBonds

pub fn len(&self) -> usize

impl FracBonds

pub fn compute(original_coords: &Coords, range: f64) -> Result<Self, Error>

Compute bonds using a uniform bond length for all types.

Numerical rounding, and distances of zero

Behavior is specified such that distances exactly equal to the interaction range are guaranteed to be included. I.e. FracBonds::from_brute_force(coords, 0.0) will produce pairs of sites that are superimposed on each other.

pub fn compute_with_meta<M: Ord>(
original_coords: &Coords,
meta: impl IntoIterator<Item = M>,
meta_range: impl FnMut(&M, &M) -> Option<f64>
) -> Result<Self, Error>

Compute bonds, using different bond lengths for different types.

meta_range must be symmetric, i.e. meta_range(a, b) == meta_range(b, a). It must also be non-negative. Neither of these properties are validated (in release mode).

Non-interacting pairs

If two types of atoms do not interact, meta_range can return None to guarantee that no bonds between these types are included in the output.

pub fn to_cart_bonds(&self, coords: &Coords) -> CartBonds

pub fn num_atoms_per_cell(&self) -> usize