Function rsp2_structure::layer::find_layers

pub fn find_layers(
    coords: &Coords, 
    miller: V3<i32>, 
    cart_threshold: f64
) -> Result<Layers, Error>

Determine layers in a structure, numbered from zero.

This finds “layers” defined as groups of atoms isolated from other groups by at least some minimum distance projected along a normal vector.

The normal vector must always point along a valid periodic plane normal; this ensures that no discontinuities arise in the coordinate at the periodic boundaries. miller determines the direction of this normal. There are two equivalent ways to perceive this:

miller is the Miller index of planes parallel to the layers (assuming they are flat).

miller is the smallest integer index of a reciprocal lattice vector that is parallel to (and positively oriented with) the normal.

Miller indices of gcd > 1 are forbidden.

Layers produced by this method will satisfy the property that all gaps are > 0 and <= the periodic length of the normal axis.