Struct rsp2_structure::Coords

pub struct Coords { /* fields omitted */ }

Pairs CoordsKind together with their Lattice.

In contrast to CoordsKind, this type is capable of providing both fractional and cartesian representations, as it contains all of the necessary information to convert between them.

Site metadata

tl;dr: You don’t need it, and Coords (along with every API it touches) is better off for not having it.

rsp2 used to have a Structure<M> type that additionally held a Vec<M> for site metadata. In the end, this was finally removed because it was found to cause more problems than it solves. Many kinds of site data couldn’t be shoehorned into the metadata because it led to unnatural representations (such as Vec<Option<L>> for something that is Option<Vec<L>> in spirit), and for this reason most code just used Structure<Element> or Structure<()> (which eventually led to the introduction of Coords).

A Structure<M> type that holds an M (for e.g. M = (Vec<Element>, Vec<Mass>)) is possible, but also an obscene design effort r.e. ownership issues, if you want something suitable for use in low-level computational code. (which Coords aims to be).

A Structure that just holds Vec<Element> solves nothing yet would demand support from all Coord-based APIs.

While rsp2 has no standard solution to this problem, it is recommended that low-level code (computations, file IO) simply take slices of the data it needs and produce Vecs. The highest-level application code in rsp2 uses HLists of Rc<[T]> with just a small number of utility extension traits to meet almost all of its needs.

(or at least, such was true when I wrote this)

Implementations

impl Coords

Construction

pub fn new(lattice: Lattice, coords: CoordsKind) -> Self

Create coordinates.

pub fn from_molecule(carts: &[V3], vacuum_sep: f64) -> Self

Create an isolated molecule, using a cubic cell with vacuum_sep padding between images on each axis.

impl Coords

Simple data access

pub fn len(&self) -> usize

Alias for num_atoms.

pub fn num_atoms(&self) -> usize

pub fn lattice(&self) -> &Lattice

impl Coords

Functions for rescaling the structure.

These functions preserve fractional position while changing the lattice.

pub fn set_lattice(&mut self, lattice: &Lattice)

pub fn scale_vecs(&mut self, scale: &[f64; 3])

Individually multiply each lattice vector by a scale factor

impl Coords

Reading coordinates

pub fn to_carts(&self) -> Vec<V3>

pub fn to_fracs(&self) -> Vec<V3>

pub fn as_carts_cached(&self) -> Option<&[V3]>

pub fn as_fracs_cached(&self) -> Option<&[V3]>

impl Coords

Modifying coordinates

pub fn carts_mut(&mut self) -> &mut [V3]

Mutably borrow the cartesian coordinates.

This forcibly clears any cached fractional data.

pub fn fracs_mut(&mut self) -> &mut [V3]

Mutably borrow the fractional coordinates.

This forcibly clears any cached cartesian data.

pub fn set_coords(&mut self, coords: CoordsKind)

Replace the coordinates in the structure.

Panics

Panics if the length does not match. (this restriction is a holdover from earlier designs, but is now fairly arbitrary, and may be lifted in the future.)

pub fn set_carts(&mut self, carts: Vec<V3>)

pub fn set_fracs(&mut self, fracs: Vec<V3>)

impl Coords

Populating caches

pub fn ensure_carts(&mut self)

Ensures that the cartesian coordinates are cached if they aren’t already.

It is unspecified whether fractional coordinates remain cached after this call.

pub fn ensure_fracs(&mut self)

Ensures that the fractional coordinates are cached if they aren’t already.

It is unspecified whether cartesian coordinates remain cached after this call.

impl Coords

Chainable modification

pub fn with_coords(&self, coords: CoordsKind) -> Self

Construct a new Coords with the same lattice. Length need not match.

pub fn with_carts(&self, carts: Vec<V3>) -> Self

pub fn with_fracs(&self, fracs: Vec<V3>) -> Self

impl Coords

Spatial operations

pub fn translate_frac(&mut self, v: &V3)

pub fn translate_cart(&mut self, v: &V3)

pub fn transform(&mut self, m: &M33)

Applies a cartesian transformation matrix.

This will keep fractional positions fixed by rotating the lattice instead.

impl Coords

Transformations between equivalent cells

pub fn apply_unimodular(&mut self, m: &M33<i32>)

Take a linear combination of the lattice vectors to produce an identical structure with a different choice of primitive cell.

Each row of the input are integer coefficients of a lattice vector to be used in the output. The absolute value of the matrix determinant must be 1.

Cartesian coordinates are preserved.

Panics

Panics if abs(det(m)) != 1.

pub fn use_equivalent_cell(
&mut self,
tol: f64,
target_lattice: &Lattice
) -> Result<(), NonEquivalentLattice>

Produce an identical structure (up to precision loss) represented in terms of a different unit cell.

The new cell is required to be equivalent to the original cell. Otherwise, it fails with NonEquivalentLattice.

Numerically speaking, the input cell will not be used exactly; instead, the new lattice is recomputed as a linear combination of the original lattice vectors. The expectation is that this causes less of a disturbance to the positions of sites which are distant to the origin, in the case where one wishes to reduce the sites into the new unit cell afterwards.

pub fn reduce_positions(&mut self)

Reduces all fractional coordinates into [0.0, 1.0).

impl Coords

Structure matching

pub fn perm_to_match(
&self,
target: &Coords,
tol: f64
) -> Result<Perm, PositionMatchError>

Get the permutation that, when applied to this structure, makes it match the target (when reduced under the lattice).

The existing implementation is pretty strict. I mean, structure matching is kind of a tricky problem.

pub fn set_difference_indices(&self, other: &Coords, tol: f64) -> Missing

Get indices of sites that are only in one structure, but not another.

The structures must have matching lattices.

FIXME: The current implementation has a known flaw that it will become significantly slower (jumping up from from O(n log n) to O(n^2) complexity) if the second structure is missing anything. I have an inkling how to fix it but I’m too lazy to benchmark it.

impl Coords

pub fn check_same_cell_and_order(
&self,
other: &Coords,
cart_tol: f64
) -> Result<(), Error>

An approximate equality check which:

Panics on mismatched length.

Does not permit uniform translations or rotations.

Does not permit equivalent cells. (lattice matrices must be approximately equal)

Does not permit reordering of sites.

Permits different periodic images of the same site.

May erroneously fail on heavily skewed unit cells where a significant portion of the voronoi cell lies outside of the 27 unit cells around the origin.