Struct rsp2_structure::Lattice

pub struct Lattice { /* fields omitted */ }

Defines a vector basis for periodic boundary conditions in three dimensions.

This type currently has two conflated roles:

A linear transformation between “fractional” data to “cartesian” data. See CoordsKind for definitions.

A 3x3 matrix with a precomputed inverse. (you might see this usage referred to as “abuse”…)

Implementations

impl Lattice

pub fn new(matrix: &M33) -> Self

Create a lattice from a matrix where the rows are lattice vectors.

pub fn from_vectors(vectors: &[V3; 3]) -> Self

pub fn reciprocal(&self) -> Self

Get the reciprocal lattice.

This is defined as the inverse transpose. There is no 2 PI factor. It is the lattice that transforms between reciprocal space fractional and reciprocal space Euclidean coordinates.

Notice that partial derivatives with respect to a coordinate system also transform like the reciprocal of that coordinate system.

pub fn matrix(&self) -> &M33

Matrix where lattice vectors are rows.

pub fn inverse_matrix(&self) -> &M33

Get the (precomputed) inverse of the matrix where lattice vectors are rows.

pub fn vectors(&self) -> &[V3; 3]

pub fn norms(&self) -> [f64; 3]

pub fn sqnorms(&self) -> [f64; 3]

pub fn volume(&self) -> f64

Get the (positive) volume of the lattice cell.

pub fn transformed_by(&self, m: &M33) -> Lattice

Apply a cartesian transformation to the lattice.

pub fn linear_combination(&self, coeffs: &M33<i32>) -> Lattice

Take an integer linear combination of the lattice vectors.

pub fn plane_normal(&self, miller: V3<i32>) -> V3

Get the normal to the family of planes with a given Miller index.

Miller indices with gcd > 1 are permitted and are interpreted according to the documentation in the crate::miller module. They should produce the same value as a primitive vector, numerical errors not withstanding.

Panics

Panics on a Miller index of [0, 0, 0].

pub fn plane_spacing(&self, miller: V3<i32>) -> f64

Get the interplanar spacing of the family of planes specified by a Miller index.

Miller indices with gcd > 1 are permitted and are interpreted according to the documentation in the crate::miller module. They should produce a spacing that is 1 / gcd times the typical spacing.

Panics

Panics on a Miller index of [0, 0, 0].

pub fn is_equivalent_to(&self, tol: f64, other: &Lattice) -> bool

Test if two Lattices represent the same lattice, in the mathematical sense. This is to say that they each generate the same infinite set of displacement vectors.

The row-based matrices A and B both represent the same Bravais lattice if and only if A B^-1 is unimodular.

tol is an absolute tolerance used to test integerness of a float.

pub fn rotate_to_lower_triangular(&self) -> Lattice

Rotate into a form where the first vector points entirely along x, and the second vector is in the xy-plane.

This is the form required by LAMMPS and V_Sim. This form makes the lattice matrix lower triangular in RSP2’s convention.

If the input cell is left-handed, the output will also be left-handed; however it is not specified which of the 3 diagonal elements will be negative.

impl Lattice

Helper constructors

pub fn eye() -> Self

The identity lattice.

pub fn diagonal(diag: &[f64; 3]) -> Self

pub fn cubic(a: f64) -> Self

A cubic lattice ((a, a, a), (90, 90, 90))

pub fn orthorhombic(a: f64, b: f64, c: f64) -> Self

An orthorhombic lattice ((a, b, c), (90, 90, 90))