Struct rsp2_structure::supercell::SupercellToken

pub struct SupercellToken { /* fields omitted */ }

Contains enough information to deconstruct a supercell produced by this library.

The order of the atoms in the supercell is unspecified and may change with the needs of rsp2. You must use the methods on this type if you need to convert between index representations, or to work in the subgroup of cell images.

In general, this represents an arbitrary supercell matrix (a 3x3 matrix of integers with nonzero determinant), though most supercells are diagonal.

It provides a variety of methods for converting between various forms of indices, summarized below along with the terms that frequently appear in method names:

primitive_atom: usize - index of a site in the original structure. 0 <= primitive_atom < num_primitive_atoms.

atom: usize - index of a site in the superstructure. 0 <= atom < num_supercell_atoms.

cell: [u32; 3] - the index of an image of the primitive structure. 0 <= cell[k] < periods[k], Like site indices, these are considered to be unique identifiers, and are expected to be in bounds. The vector difference of two cells is a meaningless quantity.

lattice_point: V3<i32> - a more mathematically-oriented form of cell. It is the integer coordinates of the lattice vector that were added to a primitive site to produce a given supercell site. In method outputs, this will be in the range 0 <= lattice_point[k] - offset[k] < periods[k]. In method inputs, this will generally be allowed to be any arbitrary integer vector.

Implementations

impl SupercellToken

pub fn periods(&self) -> [u32; 3]

👎 Deprecated:
Use as_diagonal() instead if you require a diagonal cell. If you simply want to iterate over cells, consider iterating over lattice points instead. If you really need this, and are writing code that does properly support non-diagonal cells, use iteration_periods().

pub fn iteration_periods(&self) -> [u32; 3]

Get the number of cells along each axis, for the purpose of manually iterating over **cell**s or translational symmetries. (and only these purposes!)

This is defined even for non-diagonal supercell matrices; it is the absolute values of the diagonal of the Hermite Normal Form of the supercell matrix. (note: it is unspecified whether the upper triangular or lower triangular HNF is used, so do not compute this yourself!)

Important: Do not use this to communicate information about the supercell to other software. Use Self::matrix or Self::as_diagonal.

pub fn matrix(&self) -> M33<i32>

Get the supercell matrix.

pub fn as_diagonal(&self) -> Option<[u32; 3]>

Get the periods of the supercell, if and only if it is diagonal.

This exists for interfacing with other software that does not support arbitrary supercell matrices.

pub fn num_cells(&self) -> usize

The number of images taken of the unit cell contained in the supercell.

pub fn num_primitive_atoms(&self) -> usize

pub fn num_supercell_atoms(&self) -> usize

pub fn replicate<M>(&self, vec: &[M]) -> Vec<M> where
M: Clone,

Takes data for each atom of the primitive cell and expands it to the size of the supercell.

pub fn replicate_with<M, M2, F>(&self, vec: &[M], make_meta: F) -> Vec<M2> where
F: FnMut(&M, [u32; 3]) -> M2,

pub fn deconstruct(
&self,
validation_radius: f64,
coords: Coords
) -> Result<Coords, BigDisplacement>

Recover a primitive cell by averaging positions from a supercell.

May fail if any of the following have occurred since the token was created:

Addition or deletion of atoms

Reordering of atoms

Wrapping of positions (FIXME unnecessary limitation)

Images of an atom did not move by equal amounts (within validation_radius)

pub fn collapse<M>(&self, meta: Vec<M>) -> Vec<M>

The equivalent of deconstruct for site metadata.

Uses metadata from the first image of each atom. See collapse_with for more control.

pub fn collapse_with<M, M2, F>(&self, meta: Vec<M>, fold_meta: F) -> Vec<M2> where
F: FnMut(OwnedMetas<'_, M>, &[[u32; 3]]) -> M2,

The equivalent of deconstruct for site metadata.

fold_meta is called on the images of each primitive atom, along with their cell indices in a parallel slice. The primitive atoms are done in an arbitrary order.

pub fn try_collapse_with<E, M, M2, F>(
&self,
meta: Vec<M>,
fold_meta: F
) -> Result<Vec<M2>, E> where
F: FnMut(OwnedMetas<'_, M>, &[[u32; 3]]) -> Result<M2, E>,

fold_meta is called on the images of each primitive atom, along with their cell indices in a parallel slice. The primitive atoms are done in an arbitrary order.

Variant of collapse_with for fallible functions.

pub fn center_cell(&self) -> [u32; 3]

Get the cell index of the centermost cell.

Ties on even-dimensioned axes are broken towards zero.

If you need this, then you’ve probably written some dumb algorithm that only works with odd-dimensioned supercells that are at least 3x3x3 large. You should fix that.

pub fn lattice_point_from_cell(&self, cell: [u32; 3]) -> V3<i32>

Note: The cell must be in bounds.

pub fn lattice_point_from_cell_unchecked(&self, cell: [u32; 3]) -> V3<i32>

pub fn cell_from_lattice_point(&self, v: V3<i32>) -> [u32; 3]

Note: The lattice point is wrapped into the supercell.

pub fn atom_from_cell(&self, prim: usize, cell: [u32; 3]) -> usize

Note: The cell must be in bounds.

pub fn atom_from_lattice_point(
&self,
prim: usize,
lattice_point: V3<i32>
) -> usize

Note: The lattice point is wrapped into the supercell.

pub fn atom_from_cell_unchecked(&self, prim: usize, cell: [u32; 3]) -> usize

pub fn atom_cells(&self) -> Vec<[u32; 3]>

Defines which image of the primitive cell each atom in the supercell belongs to.

Each image of the primitive cell is represented by the integral coordinates of the lattice vector used to produce it. It is guaranteed that the output will contain self.num_cells() unique values, and that each one will appear exactly self.num_primitive_atoms() times.

pub fn atom_lattice_points(&self) -> Vec<V3<i32>>

Gives coefficients of the lattice vectors that were added to each atom in the supercell.

This is like cell_indices, but gives the indices as signed integer offsets. The offset from building the supercell affects this, so that e.g. centered supercells have a signed index of [0, 0, 0] assigned to atoms in the centermost unit cell.

pub fn atom_primitive_atoms(&self) -> Vec<usize>

Defines the primitive site corresponding to each supercell site.

pub fn lattice_point_translation_deperm(&self, lattice_point: V3<i32>) -> Perm

Get a depermutation representing translation by a primitive cell lattice point.

Please see conventions.md for an explanation of depermutations.